add jwt security for every connection. and check if user correspond to the token
