+++ src/notifications/notification.gateway.ts
@@ -7,11 +7,10 @@ import {
 } from '@nestjs/websockets';
 import { Server, Socket } from 'socket.io';
 import { User, Notification as NotificationT } from '@prisma/client';
-import {
-  NotificationService,
-  RealTimeNotification,
-} from './notification.service';
-import { Logger } from '@nestjs/common';
+import { RealTimeNotification } from './notification.service';
+import { Logger, UnauthorizedException } from '@nestjs/common';
+import { JwtPayload } from '../auth/auth.service';
+import { JwtService } from '@nestjs/jwt';
 
 @WebSocketGateway({ namespace: 'notification' })
 export class NotificationGateway
@@ -20,7 +19,23 @@ export class NotificationGateway
   private readonly logger = new Logger(NotificationGateway.name);
   @WebSocketServer() server: Server;
 
-  constructor(private service: NotificationService) {}
+  constructor(private jwtService: JwtService) {}
+
+  async canConnect(token: string, userId: number): Promise<boolean> {
+    try {
+      const isValid = this.jwtService.verify(token) as JwtPayload;
+      if (!isValid) {
+        throw new UnauthorizedException('Invalid credentials');
+      }
+      // check if the user is the same as the one in the token
+      if (isValid.sub.userId !== userId) {
+        throw new UnauthorizedException('Invalid credentials');
+      }
+      return true;
+    } catch (e) {
+      return false;
+    }
+  }
 
   async handleConnection(client: Socket, ...args: any[]) {
     try {
@@ -30,7 +45,7 @@ export class NotificationGateway
       if (!token) throw new Error('Token is required');
       const id = Number(userId) ?? 0;
       // check if jwt token is valid and if user is assign that jwt token
-      await this.service.canConnect(token, id);
+      await this.canConnect(token, id);
       this.logger.log(`Client connected on notification: ${client.id}`);
     } catch (e) {
       client.emit('failedToConnect', e.message);
